[Misc] Fix some truly terrible design
